Fix build failure
Previously glibc-2.21-r2.apk was retrieved from a URL which
is now no longer functioning.  This change now retrieves the
APK from the upstream Github project releases directly.

The top level Makefile was also referencing nomad-agent, but
the folder structure is actually nomad-client.  This changes
the Makefile to point to the correct location.

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>